+++ src/apps/weblib/interfacejs.h
@@ -3,7 +3,6 @@
 
 extern "C" {
 extern void openUrl(const char *);
-extern void setCursor(const void *target, const char *cursor);
 extern void callLater(void (*)(void *), void *, int later);
 }