+++ src/chart/ui/chart.h
@@ -4,7 +4,6 @@
 #include <optional>
 
 #include "base/gui/pointer.h"
-#include "base/gui/scheduler.h"
 #include "base/gui/widget.h"
 #include "base/util/eventdispatcher.h"
 #include "chart/main/chart.h"
@@ -21,7 +20,7 @@ public:
 	    doSetCursor;
 	std::function<void(const std::string &)> openUrl;
 
-	explicit ChartWidget(GUI::Scheduler &scheduler);
+	explicit ChartWidget();
 	~ChartWidget() override;
 
 	void onPointerDown(const std::shared_ptr<Gfx::ICanvas> &,
@@ -54,22 +53,16 @@ public:
 
 private:
 	Chart chart;
-	GUI::PointerEvent pointerEvent;
-	std::reference_wrapper<GUI::Scheduler> scheduler;
 	Util::EventDispatcher::event_ptr onClick;
 	Util::EventDispatcher::event_ptr onPointerMoveEvent;
-	Util::EventDispatcher::event_ptr onPointerOnEvent;
 	Util::EventDispatcher::event_ptr onWheelEvent;
 	Util::EventDispatcher::event_ptr onPointerDownEvent;
 	Util::EventDispatcher::event_ptr onPointerUpEvent;
-	bool unprocessedPointerMove{};
-	bool unprocessedPointerLeave{};
-	std::optional<int64_t> trackedMarkerId;
-	std::optional<int64_t> reportedMarkerId;
+	Util::EventDispatcher::event_ptr onPointerLeaveEvent;
 	bool needUpdate{true};
 
-	void updateCursor(const std::shared_ptr<Gfx::ICanvas> &);
-	void trackMarker();
+	void updateCursor(const std::shared_ptr<Gfx::ICanvas> &,
+	    const Geom::Point &pos);
 	const Gen::Marker *getMarkerAt(const Geom::Point &pos);
 };